% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\name{ab}
\alias{ab}
\title{Show first rows and columns of a matrix or dataframe.}
\usage{
ab(X, k = min(nrow(X), ncol(X), 10), l = k)
}
\arguments{
\item{X}{matrix or dataframe.}

\item{k}{scalar - number of rows to show.}

\item{l}{scalar - number of columns to show.}
}
\value{
first \code{k} rows and \code{l} columns of \code{X}
}
\description{
Show first rows and columns of a matrix or dataframe.
}
\examples{
X <- matrix( 1:210, ncol = 15)
ab(X)
}
