% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrays.R
\name{add_arrays}
\alias{add_arrays}
\title{Add Arrays}
\usage{
add_arrays(A, B, d, operation = "addition")
}
\arguments{
\item{A}{An array.}

\item{B}{An array or vector.}

\item{d}{A vector, indicating the dimensions of A over which we want to add B.}

\item{operation}{The operation to be applied - by default addition, but multiplication is also available.}
}
\value{
The array resulting from adding array \code{B} to each layer of array \code{A} across specified dimensions \code{d}.
}
\description{
Add arrays of different sizes across chosen dimensions of the larger array.
}
\examples{
A <- array( 1:24, dim = c(2,3,4) )
B <- matrix( 1:12, nrow = 3 )
add_arrays( A, B, d = 1 )
add_arrays( A, B, d = 1, operation = "multiplication" )

}
