% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_min.r
\name{arg_min}
\alias{arg_min}
\title{Arg min for monotonically increasing functions.}
\usage{
arg_min(f, v, l, u, p)
}
\arguments{
\item{f}{function}

\item{v}{value}

\item{l}{initial lower bound}

\item{u}{initial upper bound}

\item{p}{precision}
}
\value{
\code{a}, a scalar such that \code{f(a) >= v} and \code{f(a-p) < v}, along with \code{fa = f(a)}.
}
\description{
Function to find minimum argument for which function is at least as big as some value v, within some accepted precision.
}
\examples{
sam <- sort( sample( 1:100, 10 ) )
fn <- function( x ){ if( x >= sam[1] ){ sam[sum( x >= sam )] }else{ 0 } }
arg_min( fn, v = 33, l = 0, u = 100, p = 10 )
arg_min( fn, v = 33, l = 0, u = 100, p = 0.1 )
# In the above example we know of course that the answer should be
# minimum of sam values that are greater than v.
}
