% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayListConversion.R
\name{ArrayListConversion}
\alias{ArrayListConversion}
\title{Array List Conversion}
\usage{
ArrayListConversion(X, rows = FALSE, levels = NULL)
}
\arguments{
\item{X}{A list of vectors/matrices or 2/3 dimensional array to be converted.}

\item{rows}{Option only for converting a matrix to a list of vectors which specifies whether the vectors in the list
are the rows (\code{TRUE}) or columns (\code{FALSE}, default) of the matrix.}

\item{levels}{Option for converting to a list - the names of the items of the list to have.}
}
\value{
A 2/3 dimensional array or list, depending on \code{X}.
}
\description{
Convert lists of vectors or arrays (of dimension d-1) to an array of dimension d, with the elements of the list indicated by the value of the dth dimension.

Also, convert a matrix to a list of vectors, either by row or column.
}
\examples{
X <- matrix( 1:12, ncol = 3 )
ArrayListConversion( X, levels = c("cat", "dog", "rabbit") )
Y <- matrix( 1:6, ncol = 2)
Z <- matrix( 7:12, ncol = 2)
L <- list( "cat" = Y, "dog" = Z )
ArrayListConversion( L )
}
