% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSort.R
\name{colSort}
\alias{colSort}
\title{Sort Each Column of Matrix}
\usage{
colSort(X, decreasing = FALSE, na.last = TRUE)
}
\arguments{
\item{X}{a matrix.}

\item{decreasing}{logical.  Should the sort order be increasing or decreasing?}

\item{na.last}{for controlling the treatment of \code{NA}s.
    If \code{TRUE}, missing values in the data are put last; if
    \code{FALSE}, they are put first; if \code{NA}, they are removed
    (see \sQuote{Note}.)}
}
\value{
a matrix with each column containing the sorted elements of the columns in X.
}
\description{
Sort Each Column of Matrix
}
\examples{
X <- matrix( sample( 50, 20 ), ncol = 4 )
colSort( X )
colSort( X, decreasing = TRUE )
}
\seealso{
\code{\link[base]{order}}
}
