% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareAndAdd.R
\name{CompareAndAdd}
\alias{CompareAndAdd}
\title{Compare and Add Variance Arrays and Vectors}
\usage{
CompareAndAdd(A, B)
}
\arguments{
\item{A}{array (of dimension 2 or 3) or vector.}

\item{B}{array (of dimension 2 or 3) or vector.}
}
\value{
Array resulting from the appropriate addition between A and B.
}
\description{
Check whether the dimension (up to 3) of two objects are the same - then perform appropriate addition of the two.

If one is an array and the other a matrix, the matrix is added to each layer of the array.  Otherwise, addition happens as if the "+" operator had been used between the two.

This function needs amending when we recall what the intention of said function was (adding variance arrays/matrices).  Think it may be good if there are more checks involved.  Also, do the first two dimensions (of an array) need to be equal?   If the answer if yes, then the example needs changing and this needs to be checked for.
}
\examples{
A <- array( 1:24, dim = c( 2,3,4 ) )
B <- matrix( 1:6, nrow=2 )
CompareAndAdd( A, B )
}
