% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertDegreesRadians.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert Degrees to Radians}
\usage{
deg2rad(d)
}
\arguments{
\item{d}{angle in degrees}
}
\value{
angle in radians
}
\description{
Convert Degrees to Radians
}
\examples{
deg2rad( 60 )
}
