% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate Euclidean distance between two vectors}
\usage{
euclidean_distance(x1, x2)
}
\arguments{
\item{x1}{vector}

\item{x2}{vector}
}
\value{
scalar output with the distance between the two vectors.
}
\description{
Calculate Euclidean distance between two vectors
}
\examples{
a <- c( 3, 5, 6.2 )
b <- c( 1, 9, 3.1 )
euclidean_distance( a, b )
}
