% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtNames.r
\name{ExtNames}
\alias{ExtNames}
\title{ExtNames}
\usage{
ExtNames(L, pos = 1)
}
\arguments{
\item{L}{List of objects with names which we wish to extract.}

\item{pos}{Position for setting where the objects will be extracted to, in terms of environment.}
}
\value{
Objects, to the specified environment.
}
\description{
ExtNames
}
\examples{
L <- list( "rabbit" = 3, "BigMatrix" = matrix( runif( 100 ), ncol = 10 ) )
L$rabbit
ExtNames( L )
rabbit
}
