% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullLinearModel.R
\name{FullLinearModel}
\alias{FullLinearModel}
\title{Full Linear Model Generation}
\usage{
FullLinearModel(X, y = rep(1, nrow(X)), order = 1)
}
\arguments{
\item{X}{dataframe (not a matrix).}

\item{y}{(optional) vector of responses}

\item{order}{order of the polynomial fit}
}
\value{
Full linear model for y given X.
}
\description{
Generate the full first, second or third order linear model for a given dataframe, X.
Whilst the actual models can be generated in this way, this function is mostly a useful way for extracting model formulae.
}
\examples{
data( USArrests )
FullLinearModel( USArrests )
FullLinearModel( USArrests, order = 2 )
FullLinearModel( USArrests, order = 3 )
FullLinearModel( USArrests[,c("UrbanPop", "Assault")], y = USArrests[,"Murder"], order = 2 )
}
