% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvertMatrices.R
\name{InvertMatrices}
\alias{InvertMatrices}
\title{Invert Each Matrix in an Array}
\usage{
InvertMatrices(V)
}
\arguments{
\item{V}{a 3-dimensional array}
}
\value{
an array, each matrix of which is the inversion of the matrices in V.
}
\description{
Invert each of the matrices in an array and return an array of inverted matrices.  Note that if some of the matrices can't be inverted then a matrix of "NA"'s is put in it's place.
}
\examples{
V <- array( stats::rnorm( 18, 0, 0.1) + rep( c( diag( 3 ) ), 2 ), dim = c(3,3,2) )
# ensure that in this example both matrices will be invertible
InvertMatrices( V )
}
