% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixMM.R
\name{MatrixMM}
\alias{MatrixMM}
\title{Matrix MM}
\usage{
MatrixMM(X, maximum = TRUE, column = TRUE)
}
\arguments{
\item{X}{a matrix or a dataframe.}

\item{maximum}{Take the maximum (\code{TRUE}) or minimum (\code{FALSE})...}

\item{column}{...of each column (\code{TRUE}) or row (\code{FALSE})}
}
\value{
vector of maximum or minimum values of the columns or rows of X.
}
\description{
Take the maximum or minimum of each column or row of a matrix.
}
\examples{
X <- matrix( runif( 6 ), ncol = 2 )
MatrixMM( X )
Y <- data.frame( X, row.names = c( "cat", "dog", "rabbit" ) )
MatrixMM( Y )
}
