% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinDist.R
\name{MinDist}
\alias{MinDist}
\title{MinDist}
\usage{
MinDist(X)
}
\arguments{
\item{X}{matrix or dataframe}
}
\value{
minimum distance between any two of the points in (given by the rows of) X
}
\description{
Computes the minimum distance between the points (rows) of a matrix.
}
\examples{
X <- matrix( runif( 12 ), ncol = 3 )
MinDist( X )
Y <- data.frame( X, row.names = c( "cat", "dog", "rabbit", "mouse" ) )
MinDist( Y )
}
