% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLH.R
\name{MLH}
\alias{MLH}
\title{Maximin Latin Hypercube}
\usage{
MLH(
  m,
  d = 2,
  niter = 1000,
  rect = "MinusoneOne",
  plot_best_iterations = TRUE,
  print_best_iterations = TRUE
)
}
\arguments{
\item{m}{number of points in the Latin hypercube.}

\item{d}{number of dimensions over which the Latin hypercube is required.
This can also be implicitly specified as the number of rows of \code{rect}.}

\item{niter}{Number of Latin hypercubes to sample and out of which the one
with maximum minimum distance between any two points will be chosen.}

\item{rect}{a matrix with \code{d} rows and \code{2} columns indicating the ranges over which the parameters
should be chosen.  Can also be specified by the string \code{"MinusoneOne"} if they are [-1,1] (default) or
\code{"ZeroOne"} if they are all [0,1].}

\item{plot_best_iterations}{should the current best iterations be plotted as they are found?}

\item{print_best_iterations}{should the iteration number of the best iterations be printed as they are found?}
}
\value{
\item{Dx_best}{best design according to the Maximin criterion.}
\item{c_D_best}{iteration on which the best design was generated.}
}
\description{
Generate a Maximin Latin hypercube design.  Samples many Latin hypercubes (\code{niter}) and selects
the one with maximum minimum distance between any two of its points.
}
\examples{
MLH( m = 10, d = 3 )
}
