% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMatrix.R
\name{ModelMatrix}
\alias{ModelMatrix}
\title{Model Matrix}
\usage{
ModelMatrix(X, model)
}
\arguments{
\item{X}{dataframe (not a matrix)}

\item{model}{a model}
}
\value{
the corresponding model matrix
}
\description{
Obtain the model matrix for a given data frame and model
}
\examples{
data(USArrests)
mod <- QuickFunc::FullLinearModel( USArrests[1:40,], order = 2 )
MM <- ModelMatrix( USArrests[1:10,], mod )
MM # Can use it for part of the training set.
MN <- ModelMatrix( USArrests[41:50,], mod )
MN # Can use it for some prediction points.
}
