% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultNormSamples.R
\name{MultNormSamples}
\alias{MultNormSamples}
\title{Multiple Normal Samples}
\usage{
MultNormSamples(Means, Variances, n = 1, uncorrelated = TRUE)
}
\arguments{
\item{Means}{A k by m matrix, k-vector or m-vector of mean values.  See details for information about k and m.}

\item{Variances}{A k by k by m array, k by k matrix, k by m matrix, k-vector, m-vector or scalar.}

\item{n}{number of samples required for each mean and variance combination.}

\item{uncorrelated}{A logical indicating whether each output component should be sampled as though uncorrelated with the others.  This
massively reduces the computational time TRUE (as is default).}
}
\value{
A matrix the drawn samples (as rows) appropriately.
}
\description{
Generate multiple normal samples given a collection of means and variances.
}
\details{
This function generates m k-variate normal samples for a collection of means and variances.  These samples may be univariate or multivariate, depending on whether the \code{uncorrelated} is set to be \code{TRUE} or \code{FALSE}.
}
\examples{
mu <- matrix(c(2,4,5,-1,-1,2), nrow=3)
Sigma <- array( c(2, 0.6, -0.4, 0.6, 3, 0.8, -0.4, 0.8,
2.2, 3, 0.6, -0.4, 0.6, 2, 0.8, -0.4, 0.8, 2.1), dim = c(3,3,2) )
MultNormSamples( mu, Sigma, 4, uncorrelated = FALSE ) # Variances can be an 3D array...
MultNormSamples( mu, Sigma[,,1], 4, uncorrelated = FALSE ) # ...or Variances can be a matrix.
MultNormSamples( mu[,1], Variances = diag(Sigma[,,1]), 5 ) # Means can be a vector.
MultNormSamples( mu[,1], Variances = Sigma[1,1,1], 5 )
}
