% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultTSamples.R
\name{MultTSamples}
\alias{MultTSamples}
\title{Multiple Multivariate t-distribution Samples}
\usage{
MultTSamples(Means, Variances, n, df)
}
\arguments{
\item{Means}{A vector or matrix of mean values.  If a vector, the means are 1-dimensional given by each element.
If a matrix, the mean vectors are given by row.}

\item{Variances}{Scalar or Matrix if the variances for each sample are the same.  A vector or 3D array if they are required to be different.}

\item{n}{the size of each sample.}

\item{df}{the number of degrees of freedom.

#' @details This function generates multiple multivariate t-distribution (with \code{df} degrees of freedom) samples, each of size \code{n}, for each mean scalar/vector in \code{Means} and variance scalar/matrix (if variance the same for each Mean) or vector/array \code{Variances} (if the variance is different for each Mean).  Note that there may be ways to improve the efficiency of this function!  In addition, it should only really be used if the there is a (not uncorrelated) multivariate variance structure since the multivariate sampling function is used unless \code{Means} is a vector.}
}
\value{
A matrix containing all the drawn samples (as rows).
}
\description{
Generate multiple multivariate t-distribution samples given a collection of means and variances.
}
\examples{
mu <- matrix(c(2,4,5,-1,-1,2), nrow=2)
Sigma <- array( c(2, 0.6, -0.4, 0.6, 3, 0.8, -0.4, 0.8,
2.2, 3, 0.6, -0.4, 0.6, 2, 0.8, -0.4, 0.8, 2.1), dim = c(3,3,2) )
MultTSamples( mu, Sigma, 4, df = 10 ) # Variances can be an 3D array...
MultTSamples( mu, Sigma[,,1], 4, df = 10) # ...or Variances can be a matrix.
MultTSamples( mu[1,], Variances = diag(Sigma[,,1]), 5, df = 10 ) # Means can be a vector.
MultTSamples( mu[1,], Variances = Sigma[1,1,1], 5, df = 10 )
}
