% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_dist.R
\name{periodic_dist}
\alias{periodic_dist}
\title{Periodic Difference}
\usage{
periodic_dist(X, Y = X, p = 1)
}
\arguments{
\item{X}{a vector}

\item{Y}{a vector}

\item{p}{scalar value of period}
}
\value{
difference between the values in \code{X} and \code{Y} mod \code{p}, given as a \code{length( X )} by \code{length( Y )} matrix.
}
\description{
Calculates the difference between two values under a particular period mod value.
}
\examples{
periodic_dist( runif( 3 ), runif( 4 ) )
}
