% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDiagnostics.r
\name{Prediction_diagnostics}
\alias{Prediction_diagnostics}
\title{Prediction Diagnostics}
\usage{
Prediction_diagnostics(fx, Efx, Varfx, var_nug = 1e-10)
}
\arguments{
\item{fx}{Model/system value.}

\item{Efx}{Prediction.}

\item{Varfx}{Uncertainty variance.}

\item{var_nug}{optional variance nugget.}
}
\value{
MASPE and RMSPE
}
\description{
Obtain Mean Absolute Standardised Prediction Error (MASPE) and Root Mean Standardised Prediction Error (RMSPE) between a set of predictions \code{Efx, Varfx} and the true model or system value \code{fx}.
}
\examples{
fx <- c(3,5,6)
Efx <- c(4,3,5.75)
Varfx <- rep( 0.8, 3 )
Prediction_diagnostics( fx, Efx, Varfx )
}
