% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertDegreesRadians.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert Radians to Degrees}
\usage{
rad2deg(r)
}
\arguments{
\item{r}{angle in radians}
}
\value{
angle in degrees
}
\description{
Convert Radians to Degrees
}
\examples{
rad2deg( 1 )
}
