% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stand_Obs.R
\name{Stand_Obs}
\alias{Stand_Obs}
\title{Standardised Observation}
\usage{
Stand_Obs(x, EX, VarX, abs = TRUE)
}
\arguments{
\item{x}{Observed value, given as a vector or array.}

\item{EX}{Expectation of X, given as a vector or array which is the same dimension of \code{x}.}

\item{VarX}{Variance of X, given as a vector or arraywhich is the same dimension of \code{x}.}

\item{abs}{Should the absolute standardised observation be calculated?}
}
\value{
the standardised observations (absolute if applicable) as a vector or array the same dimension as \code{x}.
}
\description{
Calculate the standardised observation of a set of quantities.
}
\details{
standardised observation is calculated as \code{( x - EX ) / sqrt( VarX )}
}
\examples{
EX = c(3, 8, 2)
VarX = rep( 0.7, 3 )
x = c(3.82, 6.34, 1.89)
Stand_Obs( x, EX, VarX )
}
