% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_density.R
\name{uniform_density}
\alias{uniform_density}
\title{Multivariate Uniform Density (over a hypercube)}
\usage{
uniform_density(x, l, u)
}
\arguments{
\item{x}{vector or matrix}

\item{l}{vector of lower bounds}

\item{u}{vector of upper bounds}
}
\value{
density of x for the multivariate uniform distribution specified by l and u.
}
\description{
Multivariate Uniform Density (over a hypercube)
}
\examples{
x <- c(3, 4)
l <- c(2.3, 4.5)
u <- c(6, 6)
uniform_density( x, l, u )
uniform_density( x + 1, l, u )
}
